# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Saylora API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user and business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutInput"
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request OTP for password reset via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: OTP sent to email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Email not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyInput"
      responses:
        "200":
          description: OTP verified, returns reset token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid token or passwords don't match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email address with OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend email verification OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationInput"
      responses:
        "200":
          description: Verification email resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Already verified or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        whatsapp:
          type: ["string", "null"]
        language:
          type: string
        emailVerified:
          type: boolean
        businessId:
          type: ["integer", "null"]
        businessName:
          type: ["string", "null"]
        businessSlug:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        businessSubscriptionType:
          type: ["string", "null"]
        businessIsActive:
          type: boolean
        businessTrialEndsAt:
          type: ["string", "null"]
          format: date-time
        businessSubscriptionExpiresAt:
          type: ["string", "null"]
          format: date-time
      required: [id, email, name, language, emailVerified, isAdmin, businessIsActive]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [user, accessToken, refreshToken]
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [accessToken, refreshToken]
    OtpVerifyResponse:
      type: object
      properties:
        resetToken:
          type: string
        message:
          type: string
      required: [resetToken, message]
    RegisterInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 2
        whatsapp:
          type: string
        language:
          type: string
          default: "en"
        businessName:
          type: string
          minLength: 2
        categoryId:
          type: integer
        countryId:
          type: integer
        cityId:
          type: integer
        acceptTerms:
          type: boolean
      required: [email, password, name, whatsapp, businessName, categoryId, countryId, cityId, acceptTerms]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: boolean
          default: false
      required: [email, password]
    LogoutInput:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    RefreshInput:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    OtpVerifyInput:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
          minLength: 6
          maxLength: 6
      required: [email, otp]
    ResetPasswordInput:
      type: object
      properties:
        resetToken:
          type: string
        password:
          type: string
          minLength: 8
        confirmPassword:
          type: string
          minLength: 8
      required: [resetToken, password, confirmPassword]
    VerifyEmailInput:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
          minLength: 6
          maxLength: 6
      required: [email, otp]
    ResendVerificationInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
...