# lib/api-zod/src/generated/types/verifyEmailInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Saylora API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyEmailInput {
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  otp: string;
}